fix: Unresolved reference R and setSmallIcon ambiguity in MissingServices + UnifiedBlockerService
- selfcontrol/MissingServices.kt: added missing 'import com.rasel.RasFocus.R'
  and changed setSmallIcon(iconRes) to setSmallIcon(iconRes as Int) to
  resolve overload ambiguity between setSmallIcon(Icon!) and setSmallIcon(Int)
- combo/selfcontrol/MissingServices.kt: same two fixes for the combo copy
- combo/selfcontrol/UnifiedBlockerService.kt: added 'import com.rasel.RasFocus.R'
  to fix Unresolved reference R at lines 1315 and 1338

+++ app/src/main/java/com/rasel/RasFocus/selfcontrol/MissingServices.kt
@@ -8,6 +8,7 @@ import android.content.Intent
 import android.os.Build
 import android.os.IBinder
 import android.content.Context
+import com.rasel.RasFocus.R
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
@@ -180,7 +181,7 @@ class UsageNotificationService : Service() {
             Notification.Builder(this, CHANNEL_ID)
                 .setContentTitle("RasFocus Protection")
                 .setContentText(if (isProtecting) "Blocking active" else "Service inactive")
-                .setSmallIcon(iconRes)
+                .setSmallIcon(iconRes as Int)
                 .setOngoing(true)
                 .build()
         } else {
@@ -188,7 +189,7 @@ class UsageNotificationService : Service() {
             Notification.Builder(this)
                 .setContentTitle("RasFocus Protection")
                 .setContentText(if (isProtecting) "Blocking active" else "Service inactive")
-                .setSmallIcon(iconRes)
+                .setSmallIcon(iconRes as Int)
                 .build()
         }
         startForeground(NOTIF_ID, notif)
